requirements.txt
tensorflow==1.13.1
opencv_python==4.1.1.26
imutils==0.5.3
scikit_image==0.15.0
pandas==0.24.2
matplotlib==3.1.0
Keras==2.2.5
numpy==1.16.4
streamlit==0.47.2
Pillow==6.2.1
skimage==0.0
scikit_learn==0.21.3
